testing for an incrementer
