+++ Neo4jClient.Extension.Test/TestData/TestDataDiagram.cd
@@ -1,0 +1,28 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<ClassDiagram MajorVersion="1" MinorVersion="1">
+  <Class Name="Neo4jClient.Extension.Test.Cypher.Person">
+    <Position X="1.25" Y="1" Width="2" />
+    <Members>
+      <Method Name="Person" Hidden="true" />
+      <Method Name="ToString" Hidden="true" />
+    </Members>
+    <TypeIdentifier>
+      <HashCode>AIACAAAAAAAABEAEAAAAAAQAIAEAEAAAAAAEAAAAAAA=</HashCode>
+      <FileName>TestData\Entities\Person.cs</FileName>
+    </TypeIdentifier>
+    <ShowAsAssociation>
+      <Property Name="HomeAddress" />
+    </ShowAsAssociation>
+  </Class>
+  <Class Name="Neo4jClient.Extension.Test.Cypher.Address">
+    <Position X="5" Y="1.25" Width="1.5" />
+    <Members>
+      <Method Name="ToString" Hidden="true" />
+    </Members>
+    <TypeIdentifier>
+      <HashCode>AAACAAAAAAAAAAgEAAAAAAAAAAAAAAAAAAIAAAAAAAA=</HashCode>
+      <FileName>TestData\Entities\Address.cs</FileName>
+    </TypeIdentifier>
+  </Class>
+  <Font Name="Segoe UI" Size="9" />
+</ClassDiagram>
